Use Object#get instead of Kernel::open to get object from backend
This change uses the object instance directly to get the file from the
backend instead of using a presigned URL to get it. Presigning was not
necessary in this method since the URL doesn't get passed on to the
client anyway. Also, by avoiding Kernel::open, we avoid having to
manage certificate stores in both `aws-sdk` and `open-uri`.